+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Binding::AnyKey` to assign any button.
+- `Cooldown` input condition.
 
 ## [0.17.0] - 2025-08-18